web/requirements.txt
Django==2.2.9
django-extensions==2.2.5
djangorestframework==3.11
psycopg2-binary==2.8.4
Pillow==2.0.0
python-decouple==3.3
requests==2.22.0
xmltodict==0.12.0
